// frontend/src/app/api/generated/model/product-definition.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductDefinition { 
    id: string;
    code: string;
    name: string;
    validityDays: number;
    entryLimit?: number | null;
    /**
     * Current default gross price in Hungarian forints.
     */
    defaultPrice: number;
    active: boolean;
}