// frontend/src/app/api/generated/model/upsert-gym-pass-product-request.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpsertGymPassProductRequest { 
    name: string;
    validityDays: number;
    entryLimit?: number | null;
    defaultPrice: number;
}